Add monadic bindings to the procedure mode.
This adds `p <- e` syntax to the procedure mode that uses
a provied `bind` operation to execute monadic actions.
Likewise, the `return` operation automatically applies
the `pure` operation of the monad, when provided.

The syntax is still a bit rough, as monadic actions must be
explicitly bound via `_ <- e` even when the return is ignored,
and an explict `return` is requred, even if the result computed
directly by the action.  Also, it would be nice to provide
the monad in question via a submodule instead of by explicitly
naming the operations.