Avoid reusing SMT-Lib variable names across calls to `proveImp`.
This is a workaround for a bug in z3-4.8.8. The "persist-solver"
PR made it so the typechecker may call `proveImp` more than once
on the same connection, and z3-4.8.8 doesn't handle this correctly
if the same variable names are reused. This PR creates a unique
name supply per solver connection, so that all generated variable
names on the same connection are distinct.