cryptol-remote-api: Upgrade to mypy-0.812
This upgrades `mypy` to version `0.812`. This is primarily motivated by a need
to work around python-poetry/poetry#3094, which currently happens if you try to
include the SAW Python bindings in a `poetry` project as a `develop`
dependency. For the most part, this is a matter of adjusting version bounds.
In a handful of cases, we are able to remove `type: ignore` comments due to
`mypy-0.8` being smarter than `mypy-0.7`.