add a setting to control how records are displayed
For human consumers, it's usually nice to show record fields in the same
order they appear in the source (as sometimes programmers use the field
order as a communication channel to other programmers). However, for
machine consumers, sometimes it is nicer to adhere to a rigid,
predictable format. This commit adds a setting, `field-order`, to switch
between these two modes, with a backward-compatible default.