Add new declaration forms that can be parsed at the REPL.
In particular, we can now define operators using `let` in the
same way that they can be inside modules and `where`.

In addition, we allow `type` and `type constraint` definitions
at the REPL, and `infix`, `infixl` and `infixr` declarations.

Becase it is an error for an infix declaration to be stated
without its corresponding declaration, we have added the ability
to enter multiple definitions at once, separated by `;`.  This
is also useful for definition mutually recursive definitions using
`let`.