Make issue{066,093} resilient to different solvers' behavior
This adopts the same approach as in commit
3ea5e9e51c260c3bdb85da0ab7ea790bbdff81a4:
use `:set show-examples=false` in the affected examples to avoid showing the
particular data that a solver picks for examples/counterexamples. Also avoid
printing the contents of `it` for similar reasons. I have verified that the
new output works across a wide range of Z3 versions.

Fixes #1280.