Fix an -Woverlapping-patterns warning uncovered by GHC 9.2
GHC 9.2 has a more aggressive pattern-match coverage checker than in previous
versions, which reveals that the `detailedPrompt` case of `mkPrompt` is
unreachable. This is technically true, as `detailedPrompt = False`, but it
would be nice if we didn't get a warning just for keeping this code around.
This simplest solution is to just change `detailedPrompt` to be `id False`,
which defeats the coverage checker.

An alternative would be to use `GHC.Exts`' newly added `considerAccessible`
function (see https://gitlab.haskell.org/ghc/ghc/-/merge_requests/5100), but
this would require CPP.