Always embed git commit into --version output
Previously, we attempted to scrub out `git` commits from release binaries using
a `sed` invocation in the `cry` script, but as was observed in #1353, this
ended up scrubbing out `git` commits on all builds, not just release ones. We
have decided that we should just include `git` commits on all builds, so we can
fix the issue by removing the buggy use of `sed`.

Unfortunately, `cabal`'s recompilation logic is not smart enough to always
rebuild `GitRev.hs` on every new commit, even though it should. For this
reason, we make the `./cry build` command temporarily dirty the `GitRev.hs`
file so that it is forcibly rebuilt, then revert the change to `GitRev.hs` and
rebuild. It's rather heavy-handed, but it works.

Fixes #1353.