Disallow direct functor instantiation in interfaces.
The reason for this is that we don't have anywhere to place the generated
module---we could put it outside the interface, but that wont' work if
the instantiation used an imported module, and it would be inconsistent
with how other such imports work.   Instead, users need to make the
substantiation manually outside the interface.

Fixes #1485