Restrict PrimMap to only refer to actual primtives.
Previously it would contain everything in the prelude.
Mostly this was not a problem because the usual use case for this is
to find out the names of actual primitives, so having more in the mapping
is not a problem.  However, this accuracy is inconveninet, if you
want to quickly look what's actually a primitive