CI: Don't run `poetry update`
This causes `poetry` to regenerate the `poetry.lock` file with newer versions
of dependencies, which defeats the point of checking in the lock file in the
first place.

This mirrors a similar change applied to SAW's CI in
https://github.com/GaloisInc/saw-script/commit/67a25a81c4f154fd791188494d739274c7f1a236.
The use of `poetry update` isn't actively causing problems with Cryptol's CI
yet, but better to be safe than sorry.