cryptol-remote-api: Check for valid evaluation contexts
The `check`, `prove or satisfy`, and `evaluate expression` commands now check
if the type and schema of the supplied expression is valid, which can prevent
panics that would arise from evaluating invalid expressions (e.g., ones that
depend on definitions in parameterized modules). We do so by defining a
`validEvalContext` function (inspired by a function of the same name in the
Cryptol REPL) and using it where appropriate.

I discovered that there was an unused error code (see the `evalInParamMod`
function) in the remote API for this exact scenario, so I have repurposed the
error code for this kind of situation. I've also expanded its payload to
include both type parameters and definitions from the parameterized module,
much like how the Cryptol REPL's `EvalInParamMod` exception currently works.

Fixes #1623.