Implement all floating-point primitives in reference evaluator
This ensures that all of the primitives defined in `lib/Float.cry` are
supported in Cryptol's reference evaluator. Moreover, this removes the
primitive table entry for `fpIsFinite`, as `fpIsFinite` is not defined as a
primitive and therefore doesn't need special treatment.

I have added a `tests/regression/float_reference_eval.icry` test case which
ensures that the reference evaluator does the expected thing on all
floating-point operations.

Fixes #1714.