CI: Disambiguate Windows MSI installer name
Previously, this Windows MSI installer was lumped into the same artifact name
as the general Windows binary distribution (e.g.,
`cryptol-3.2.0.99-windows-2019-X64 (GHC 9.4.8)`). As of `v4` of GitHub Actions'
`upload-artifact`, however, it is an error to have two separate
`upload-artifact` steps share the same artifact name. As such, we have to
disambiguate the Windows installer artifact, which this patch accomplishes.