Fix tab completion for :t and :check
If the user types `:<CMD> ...` and `<CMD>` matches a defined command exactly,
then commit to that command's tab completion. (Previously, the tab completion
logic would consider all possible commands for which `<CMD>` is a prefix, but
this is silly when you have an exact match with an existing command.)

Fixes #1780.