Add a tcSmtFile REPL option
`tcSmtFile` is to `tcSolver` as `smtFile` is to `prover`. The `tcSmtFile`
option allows users to record typechecker-related SMT solver interactions to a
file.

This requires using a more recent `simple-smt` version to bring in the changes
from https://github.com/yav/simple-smt/pull/25 and
https://github.com/yav/simple-smt/pull/27, which are needed to plumb the
relevant information down to the `simple-smt` API.

Fixes #1782.