ChaChaPolyCryptolIETF: Return an Option in AEAD_CHACHA20_POLY1305_DECRYPT (GaloisInc/cryptol-specs#239)
(This copies over the changes made to the upstream `cryptol-specs` version of
this spec in
https://github.com/GaloisInc/cryptol-specs/commit/548ad3fd10aa1af2928c77d1d9c2298cba7a1c81
as part of GaloisInc/cryptol-specs#240. This is in pursuit of the eventual goal
of purging uses of the `(a, Bit)` Cryptol anti-pattern in favor of returning
`Option` types, as described in GaloisInc/cryptol-specs#239.)

Instead of returning `([m][8], Bit)`, where the `Bit` indicates whether the
result is valid or not, we now make the `AEAD_CHACHA20_POLY1305_DECRYPT`
function return an `Option`. We also introduce an `optFold` function to
`OptionUtils` to factor out some repeatedly used code needed to post-process
the decrypted results.

This checks off one box in GaloisInc/cryptol-specs#239.