Fix simplification rule for nested type-level exponentiation
Previously, `Cryptol.TypeCheck.SimpType.tExp` would attempt to simplify `a ^^
(x ^^ y)` to `a ^^ (x * y)` at the type level, which was blatantly wrong. I
have changed this to a similar (but actually valid) rule that simplified `(a ^^
x) ^^ y` to `a ^^ (x * y)`. (This law is normally stated with the condition
that `a != 0`, but this law even works in Cryptol when `a == 0` due to the fact
that `0 ^^ 0 == 1` in Cryptol.)

Fixes #1799.