Skip test_project.py when running in Docker
Unfortunately, this test case appears to cause various permission-related
headaches (for reasons that are difficult to reproduce outside of GitHub
Actions), and what's more, it seems to cause later tests to fail (perhaps due
to the presence of a `.cryproject` file?). See #1823 for more details.

For the sake of getting the CI to work again, I am opting to just disable this
test case when running the tests via Docker.