cryptol-remote-api/Dockerfile: Don't hardcode the user ID 1000
Starting with 24.04, the Ubuntu Docker image includes an `ubuntu` user with the
user ID 1000. This conflicts with `cryptol-remote-api`'s Docker image, which
attempts to create a separate `cryptol` user with the same ID.

The use of the user ID 1000 was introduced in
https://github.com/GaloisInc/cryptol/commit/dfae4580e322584185235f301bc8a03b6bc19a65,
but it is unclear if using that ID in particular is important. Under the
assumption that it is not, this commit removes this hard-coded ID.