cry-ffi: Add associated Parameters type to Exportable trait
This allows users to customize how values are exported in `Exportable` impls,
which allows us to define `Exportable` impls for things like `Vec`s and slices
(which require custom `Parameters` to dictate the length of the exported
value). This avoids needing to define separate top-level functions for
exporting these types.

The design here is heavily influenced by the `Arbitrary` trait in the
`proptest` crate, which also uses an associated `Parameters` type. The main
difference is that `Arbitrary` requires its `Parameters` to implement
`Default`, which `Exportable` (intentionally) does not.