CryptolTC.z3: Use SMT-LIB 2.6 syntax for declare-datatypes
Previously, `CryptolTC.z3` was using a Z3-specific syntax for the
`declare-datatypes` command that differs from the syntax prescribed by SMT-LIB
2.6. This meant that non-Z3 solvers (e.g., CVC5) could not parse the datatypes
declared in the file.

Thankfully, Z3 4.8.8 or later accept both `declare-datatypes` syntaxes. To
allow for non-Z3 solvers to be used, I have switched `CryptolTC.z3` over to the
SMT-LIB syntax. This suffices to allow alternative solvers like CVC5 to be used
as the `tcSolver`, as demonstrated in the newly added `issue1874.icry` test
case.

Fixes #1874.