better-errors: Catch timeCmd parse errors
When the timeCmd does not parse properly because of errors in its
arguments (including empty arguments), this commit catches the
parseError, prints the commands help message, then raises the error.

This allows users who don't know how to use the command to get a better
sense of what it's supposed to do.