better-errors:Change error msg constructor + msg
This commit changes the name of the function (which constructs the error
message for the specifc case when a command is not given an expression
argument when it expects one) to `invalidCommandArgument`. This function
name is more representative of the kind of errors we are dealing with
here.

This commit additionally changes the error message to say what the error
is (i.e. an EXPR arg is missing) and directing the user to use the help
command. This is a much more helpful message and avoids clutering the
error message with things that the user may not need.