Use non-simplifying apSubst in solveDerivedInst
The regular apSubst automatically recurses into simplify after
performing substitution, causing solveDerivedInst which is called from
simplifyStep to be not really just a "step" and resulting in less
precise error messages. Plus, it causes a module import cycle which is
undesirable.

Create a version of apSubst for Type which does not call simplify, and
put it in a separate module so we can get rid of the module import
cycle. This requires moving half of the things in
Cryptol.TypeCheck.Subst into a new module Cryptol.TypeCheck.Subst.Type,
but we still re-export everything in Cryptol.TypeCheck.Subst so there
should be no external impact.