refman: Update table of keywords
Add `case`, `of`, `enum`, `deriving`, and `foreign`.

Remove `as` and `hiding`, since they are allowed to be used as
identifiers. Also remove `extern`; I'm not sure what it's for, but it
doesn't seem to be used in the parser.

Change chop.hs to output 5 rows per column, since with 4 rows the table
now overflows and causes a scroll bar to appear in the html.