CI: Exclude Dependabot from actions where forks aren't permitted
Although Dependabot PRs technically run CI as though they were a branch in the
upstream repo, not a fork, in practice they do not have access to any of the
secrets required to do things such as signing binaries. As such, Dependabot PRs
effectively behave as though they were running from forks, so we should treat
Dependabot accordingly in the CI.

Fixes #1940.